使用string append时指定buffer长度

+++ services/snapshot/kernel_snapshot_manager.cpp
@@ -61,7 +61,7 @@ std::string KernelSnapshotManager::ReadKernelSnapshot()
     do {
         ret = read(snapshotFd, buffer, BUFFER_LEN - 1);
         if (ret > 0) {
-            snapshotCont.append(buffer, ret);
+            snapshotCont.append(buffer, static_cast<size_t>(ret));
         }
         if (ret < 0) {
             DFXLOGE("read snapshot filed %{public}d", errno);
